requirements.txt
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==7.0.0
torch==1.11.0
torchvision==0.12.0
wandb==0.12.16